requirements.txt
pandas~=1.5.3
Pillow~=8.4.0
seaborn~=0.12.2
matplotlib~=3.6.3
numpy~=1.24.1
setuptools~=65.5.1
